hephestus/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
requests>=2.28.2
python-dotenv>=1.0.0
pytest>=7.3.1
logging>=0.4.9
gitpython>=3.1.30